+++ extract_msg/recipient.py
@@ -12,17 +12,16 @@ import logging
 import weakref
 
 from typing import (
-        Any, Callable, Generic, List, Optional, Tuple, TYPE_CHECKING, Type,
-        TypeVar, Union
+        Any, Callable, Generic, List, Optional, TYPE_CHECKING, Type, TypeVar,
+        Union
     )
 
-from .constants import MSG_PATH
+from .constants import MSG_PATH, OVERRIDE_CLASS
 from .enums import ErrorBehavior, PropertiesType
 from .exceptions import StandardViolationError
-from .properties.prop import FixedLengthProp
 from .properties.properties_store import PropertiesStore
 from .structures.entry_id import PermanentEntryID
-from .utils import msgPathToString, verifyPropertyId, verifyType
+from .utils import msgPathToString
 
 
 if TYPE_CHECKING:
@@ -56,131 +55,6 @@ class Recipient(Generic[_RT]):
         self.__typeFlags = self.__props.getValue('0C150003', 0)
         self.__type = recipientTypeClass(0xF & self.__typeFlags)
         self.__formatted = f'{self.__name} <{self.__email}>'
-
-    def _getStream(self, filename : MSG_PATH) -> Optional[bytes]:
-        """
-        Gets a binary representation of the requested filename.
-
-        This should ALWAYS return a bytes object if it was found, otherwise
-        returns None.
-
-        :raises ReferenceError: The associated MSGFile instance has been garbage
-            collected.
-        """
-        import warnings
-        warnings.warn(':method _getStream: has been deprecated and moved to the public api. Use :method getStream: instead (remove the underscore).', DeprecationWarning)
-        return self.getStream(filename)
-
-    def _getStringStream(self, filename : MSG_PATH) -> Optional[str]:
-        """
-        Gets a string representation of the requested filename.
-
-        Rather than the full filename, you should only feed this function the
-        filename sans the type. So if the full name is "__substg1.0_001A001F",
-        the filename this function should receive should be "__substg1.0_001A".
-
-        This should ALWAYS return a string if it was found, otherwise returns
-        None.
-
-        :raises ReferenceError: The associated MSGFile instance has been garbage
-            collected.
-        """
-        import warnings
-        warnings.warn(':method _getStringStream: has been deprecated and moved to the public api. Use :method getStringStream: instead (remove the underscore).', DeprecationWarning)
-        return self.getStringStream(filename)
-
-    def _getTypedAs(self, _id : str, overrideClass = None, preserveNone : bool = True):
-        """
-        Like the other get as functions, but designed for when something
-        could be multiple types (where only one will be present). This way you
-        have no need to set the type, it will be handled for you.
-
-        :param overrideClass: Class/function to use to morph the data that was
-            read. The data will be the first argument to the class's __init__
-            function or the function itself, if that is what is provided. By
-            default, this will be completely ignored if the value was not found.
-        :param preserveNone: If true (default), causes the function to ignore
-            :param overrideClass: when the value could not be found (is None).
-            If this is changed to False, then the value will be used regardless.
-        """
-        value = self._getTypedData(_id)
-        # Check if we should be overriding the data type for this instance.
-        if overrideClass is not None:
-            if value is not None or not preserveNone:
-                value = overrideClass(value)
-
-        return value
-
-    def _getTypedData(self, _id, _type = None):
-        """
-        Gets the data for the specified id as the type that it is supposed to
-        be. :param id: MUST be a 4 digit hexadecimal string.
-
-        If you know for sure what type the data is before hand, you can specify
-        it as being one of the strings in the constant FIXED_LENGTH_PROPS_STRING
-        or VARIABLE_LENGTH_PROPS_STRING.
-
-        :raises ReferenceError: The associated MSGFile instance has been garbage
-            collected.
-        """
-        verifyPropertyId(id)
-        _id = _id.upper()
-        found, result = self._getTypedStream('__substg1.0_' + _id, _type)
-        if found:
-            return result
-        else:
-            found, result = self._getTypedProperty(_id, _type)
-            return result if found else None
-
-    def _getTypedProperty(self, propertyID : str, _type = None) -> Tuple[bool, Optional[object]]:
-        """
-        Gets the property with the specified id as the type that it is supposed
-        to be. :param id: MUST be a 4 digit hexadecimal string.
-
-        If you know for sure what type the property is before hand, you can
-        specify it as being one of the strings in the constant
-        FIXED_LENGTH_PROPS_STRING or VARIABLE_LENGTH_PROPS_STRING.
-        """
-        verifyPropertyId(propertyID)
-        if _type:
-            verifyType(_type)
-            prop = self.props.get(propertyID + _type)
-            if isinstance(prop, FixedLengthProp):
-                return True, prop.value
-            else:
-                return False, None
-        else:
-            props = self.props.getProperties(propertyID)
-            for prop in props:
-                if isinstance(prop, FixedLengthProp):
-                    return True, prop.value
-
-        return False, None
-
-    def _getTypedStream(self, filename : MSG_PATH, _type = None):
-        """
-        Gets the contents of the specified stream as the type that it is
-        supposed to be.
-
-        Rather than the full filename, you should only feed this function the
-        filename sans the type. So if the full name is "__substg1.0_001A001F",
-        the filename this function should receive should be "__substg1.0_001A".
-
-        If you know for sure what type the stream is before hand, you can
-        specify it as being one of the strings in the constant
-        FIXED_LENGTH_PROPS_STRING or VARIABLE_LENGTH_PROPS_STRING.
-
-        If you have not specified the type, the type this function returns in
-        many cases cannot be predicted. As such, when using this function it is
-        best for you to check the type that it returns. If the function returns
-        None, that means it could not find the stream specified.
-
-        :raises ReferenceError: The associated MSGFile instance has been garbage
-            collected.
-        """
-        if (msg := self.__msg()) is None:
-            raise ReferenceError('The msg file for this Recipient instance has been garbage collected.')
-        return msg._getTypedStream(self, [self.__dir, msgPathToString(filename)], True, _type)
 
     def exists(self, filename : MSG_PATH) -> bool:
         """
@@ -247,7 +121,7 @@ class Recipient(Generic[_RT]):
             raise ReferenceError('The msg file for this Recipient instance has been garbage collected.')
         return msg.getMultipleString([self.__dir, msgPathToString(filename)])
 
-    def getPropertyAs(self, propertyName : Union[int, str], overrideClass : Callable[[Any], _T]) -> Optional[_T]:
+    def getPropertyAs(self, propertyName : Union[int, str], overrideClass : OVERRIDE_CLASS[_T]) -> Optional[_T]:
         """
         Returns the property, setting the class if found.
 
@@ -324,7 +198,7 @@ class Recipient(Generic[_RT]):
             raise ReferenceError('The msg file for this Recipient instance has been garbage collected.')
         return msg.getStream([self.__dir, msgPathToString(filename)])
 
-    def getStreamAs(self, streamID : MSG_PATH, overrideClass : Callable[[Any], _T]) -> Optional[_T]:
+    def getStreamAs(self, streamID : MSG_PATH, overrideClass : OVERRIDE_CLASS[_T]) -> Optional[_T]:
         """
         Returns the specified stream, modifying it to the specified class if it
         is found.
@@ -360,7 +234,7 @@ class Recipient(Generic[_RT]):
             raise ReferenceError('The msg file for this Recipient instance has been garbage collected.')
         return msg.getStringStream([self.__dir, msgPathToString(filename)])
 
-    def getStringStreamAs(self, streamID : MSG_PATH, overrideClass : Callable[[Any], _T]) -> Optional[_T]:
+    def getStringStreamAs(self, streamID : MSG_PATH, overrideClass : OVERRIDE_CLASS[_T]) -> Optional[_T]:
         """
         Returns the specified string stream, modifying it to the specified
         class if it is found.