+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 **v0.46.0**
 * [[TeamMsgExtractor #95](https://github.com/TeamMsgExtractor/msg-extractor/issues/95)] Adjusted the `overrideEncoding` property of `MSGFile` to allow automatic encoding detection. Simply set the property to the string `"chardet"` and, assuming the `chardet` module is installed, it will analyze a number of the strings to try and form a consensus about the encoding. This will *ignore* the specified encoding *only if* if successfully detects. Otherwise it will log a warning and fall back to the default behavior.
+* Removed methods deprecated in `v0.45.0`.
 * Changed the base class of `EntryID` from no base class to `abc.ABC`.
 * Added `position` property to `EntryID` to tell how many bytes were used to create the `EntryID`.
 * Added a number of properties to `MSGFile` from \[MS-OXCMSG\].
@@ -42,6 +43,7 @@
 * Changed various data structs to allow modification and creation of new instances for writing to an MSG file.
 * changed `TZRule` to use unsigned values where applicable.
 * Changed `TZRule` to require the 14 null bytes (I noticed there is a note about outlook violating that standard and will look into it).
+* Removed unneeded function `windowsUnicode`.
 
 **v0.45.0**
 * BREAKING: Changed parsing of string multiple properties to remove the trailing null byte. This *will* cause the output of parsing them to differ.